requirements.txt
Flask
gunicorn
scikit-learn
psycopg2-binary